temporary commit

+++ src/esp32/driver/Storage_SD.cpp
@@ -1,5 +1,7 @@
 #include "Storage_SD.h"
 
+#define USE_SD
+
 const char *SDTag = "Storage_SD";
 
 Storage_SD::Storage_SD(int chipSelect) : _chipSelect(chipSelect)
@@ -12,6 +14,7 @@ Storage_SD::~Storage_SD()
 
 bool Storage_SD::init()
 {
+    #ifdef USE_SD
     while (!SD.begin(_chipSelect))
     {
         ESP_LOGE(SDTag, "SD Card Mount Failed");
@@ -19,12 +22,23 @@ bool Storage_SD::init()
     }
 
     ESP_LOGI(SDTag, "SD Card Mount Success");
+    #else
+    if (!LittleFS.begin())
+    {
+        ESP_LOGE(SDTag, "LittleFS Mount Failed");
+        return false;
+    }
+    #endif
     return true;
 }
 
-fs::File Storage_SD::openFile(const String& fileName, const char* mode)
+fs::File Storage_SD::openFile(const String &fileName, const char *mode)
 {
+    #ifdef USE_SD
     File file = SD.open(fileName, mode);
+    #else
+    fs::File file = LittleFS.open(fileName, mode);
+    #endif
     if (!file)
     {
         ESP_LOGE(SDTag, "Failed to open file for reading");
@@ -33,9 +47,13 @@ fs::File Storage_SD::openFile(const String& fileName, const char* mode)
     return file;
 }
 
-bool Storage_SD::deleteFile(const String& fileName)
+bool Storage_SD::deleteFile(const String &fileName)
 {
+    #ifdef USE_SD
     if (SD.remove(fileName))
+    #else
+    if (LittleFS.remove(fileName))
+    #endif
     {
         ESP_LOGI(SDTag, "File deleted: %s", fileName);
         return true;
@@ -49,26 +67,41 @@ bool Storage_SD::deleteFile(const String& fileName)
 
 bool Storage_SD::deleteAllFiles(const char *dir)
 {
-    File root = SD.open(dir);
+    ESP_LOGI(SDTag, "Deleting all files in %s", dir);
+    #ifdef USE_SD
+    if (!SD.begin(_chipSelect))
+    {
+        ESP_LOGE(SDTag, "SD Card not mounted");
+        return false;
+    }
+    fs::File root = SD.open(dir);
+    #else
+    fs::File root = LittleFS.open(dir);
+    #endif
+    if (!root)
+    {
+        ESP_LOGE(SDTag, "Failed to open directory: %s", dir);
+        return false;
+    }
+
     while (true)
     {
-        File entry = root.openNextFile();
-        String fileName = "/" + String(entry.name());
+        ESP_LOGD(SDTag, "Deleting files in %s", dir);
+        fs::File entry = root.openNextFile();
         if (!entry)
         {
-            // No more files
+            ESP_LOGD(SDTag, "No more files");
             break;
         }
 
+        String fileName = "/" + String(entry.name());
         if (entry.isDirectory())
         {
-            // If it's a directory, recursively delete its contents
             ESP_LOGI(SDTag, "Deleting directory: %s", fileName.c_str());
             deleteDirectory(fileName.c_str());
         }
         else
         {
-            // If it's a file, delete it
             ESP_LOGI(SDTag, "Deleting file: %s", fileName.c_str());
             SD.remove(fileName.c_str());
         }
@@ -81,7 +114,16 @@ bool Storage_SD::deleteAllFiles(const char *dir)
 
 bool Storage_SD::deleteDirectory(const char *dirname)
 {
+    #ifdef USE_SD
+    if (!SD.begin(_chipSelect))
+    {
+        ESP_LOGE(SDTag, "SD Card not mounted");
+        return false;
+    }
     File dir = SD.open(dirname);
+    #else
+    fs::File dir = LittleFS.open(dirname);
+    #endif
     while (true)
     {
         File entry = dir.openNextFile();
@@ -99,13 +141,21 @@ bool Storage_SD::deleteDirectory(const char *dirname)
         {
             // Delete file
             ESP_LOGI(SDTag, "Deleting file: %s", entry.name());
+            #ifdef USE_SD
             SD.remove(entry.name());
+            #else
+            LittleFS.remove(entry.name());
+            #endif
         }
 
         entry.close();
     }
     // After directory is empty, remove it
+    #ifdef USE_SD
     SD.rmdir(dirname);
+    #else
+    LittleFS.rmdir(dirname);
+    #endif
 
     return true;
 }
